+++ R/Hidden_Functions.R
@@ -274,6 +274,22 @@
     as.data.frame(lapply(x, function(y) { levels(y) <- Vseq; y} ))
 }
 
+.km_dist          <- function(mode, obj, frwts = NULL)     {
+  if(is.null(frwts)) return(sum(mode   != obj))
+  obj             <- as.character(obj)
+  mode            <- as.character(mode)
+  different       <- which(mode        != obj)
+  n_mode          <- 
+  n_obj           <- numeric(length(different))
+  for(i in seq_along(different))        {
+    frwt          <- frwts[[different[i]]]
+    names         <- names(frwt)
+    n_mode[i]     <- frwt[which(names  == mode[different[i]])]
+    n_obj[i]      <- frwt[which(names  ==  obj[different[i]])]
+  }
+    return(sum((n_mode + n_obj)/(n_mode * n_obj)))
+}
+
 .lab_width        <- function(x, cex = 0.67, offset = 1.5) {
     (offset + max(graphics::strwidth(x, units = "inches")) * graphics::par("mar")[1L]/graphics::par("mai")[1L]) * cex
 }
@@ -685,6 +701,12 @@
     tmp/sum(tmp)
 }
 
+.rand_mode  <- function(x) {
+  a         <- which.min(x)
+  a         <- if(length(a) > 1) sample(a, 1L) else a
+    return(a)
+}
+
 #' @importFrom matrixStats "rowSums2"
 .renorm_z         <- function(z) z/rowSums2(z)
 
@@ -808,6 +830,15 @@
     z[classification  == groups[j], j] <- 1L
   }
     return(z)
+}
+
+.update_mode      <- function(num, cluster, data, weights = NULL)  {
+  diff            <- which(cluster == num)
+  clust           <- data[diff,, drop=FALSE]
+  apply(clust, 2L, function(cat)    {
+    cat           <- if(is.null(weights)) table(cat) else tapply(weights[diff], cat, FUN=sum)
+      return(names(cat)[which.max(cat)])
+  })
 }
 
 .version_above  <- function(pkg, versi) {